+++ e-learning/app/Extensions/ManagerDataProvider.php
@@ -9,6 +9,8 @@ namespace App\Extensions;
 use \Illuminate\Contracts\Auth\UserProvider;
 use App\Extensions\Manager;
 use database\connectors\ManagerData;
+use UserAgentParser\Exception\NoResultFoundException;
+use UserAgentParser\Provider\WhichBrowser;
 class ManagerDataProvider implements UserProvider {
     private $model;
