+++ e-learning/app/Http/Controllers/ProfileController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers;
+
+
+class ProfileController extends Controller{
+    public function index(){
+        return view('profile');
+    }
+}
