+++ e-learning/resources/views/profile.blade.php
@@ -1,0 +1,15 @@
+<!-- Extends HTML structure from layouts/app.blade -->
+@extends('layouts.app')
+<style>
+
+
+
+</style>
+<Title>My Profile</Title>
+@include('inc.course.banner2')
+@include('inc.home.popup')
+<div class="container">
+
+
+
+</div>
